# apps/backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Our Haven API
  version: 0.0.0
  description: Our Haven v1 API — Parent + Provider mobile, Provider web portal,
    admin dashboard. Source of truth per ADR-0004 (§§ 1–3, 8) and ADR-0010
    (Supabase + Fly.io + Vercel platform). US-region only.
  license:
    name: Proprietary
components:
  securitySchemes:
    supabaseAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token (HS256). Verified locally with the
        project JWT secret on every request.
...
paths:
...
  /v1/providers/me/stripe-connect/summary:
    get:
      summary: Read the authenticated Provider's Stripe Connect Express account summary
      tags:
        - providers
      description: Returns the read-only Connect account state mirrored from
        `account.updated` webhooks. Includes charges_enabled / payouts_enabled /
        details_submitted, the disabled_reason if Stripe has paused the account,
        and the requirement lists (currently_due, past_due,
        pending_verification) so the UI can surface what the Provider still
        needs to fix on Stripe's side. `accountReady` is true iff both
        capabilities are enabled — this is the OH-110 gate for verification
        activation + appearing in search.
      security:
        - supabaseAccessToken: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasAccount:
                    type: boolean
                  stripeAccountId:
                    anyOf:
                      - type: string
                      - type: "null"
                  chargesEnabled:
                    type: boolean
                  payoutsEnabled:
                    type: boolean
                  detailsSubmitted:
                    type: boolean
                  disabledReason:
                    anyOf:
                      - type: string
                      - type: "null"
                  accountReady:
                    type: boolean
                  accountReadyAt:
                    anyOf:
                      - type: string
                        format: date-time
                        pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                      - type: "null"
                  requirementsCurrentlyDue:
                    type: array
                    items:
                      type: string
                  requirementsPastDue:
                    type: array
                    items:
                      type: string
                  requirementsPendingVerification:
                    type: array
                    items:
                      type: string
                  lastWebhookAt:
                    anyOf:
                      - type: string
                        format: date-time
                        pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                      - type: "null"
                required:
                  - hasAccount
                  - stripeAccountId
                  - chargesEnabled
                  - payoutsEnabled
                  - detailsSubmitted
                  - disabledReason
                  - accountReady
                  - accountReadyAt
                  - requirementsCurrentlyDue
                  - requirementsPastDue
                  - requirementsPendingVerification
                  - lastWebhookAt
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
  /v1/providers/me/stripe-connect/onboarding-link:
    post:
      summary: Create / reuse a Stripe Connect Express account and return a hosted
        onboarding URL
      tags:
        - providers
      description: "Idempotent: if the Provider already has a
        `provider_connect_accounts` row with a `stripe_account_id`, reuses it;
        otherwise creates a fresh Express account (type=express, country=US,
        capabilities=card_payments+transfers, business_type=individual) and
        stamps it onto the row. Returns a freshly-issued account-onboarding link
        the client should redirect to. Stripe redirects back to
        `STRIPE_CONNECT_RETURN_URL` on completion (the verification page picks
        up the change via the webhook + summary refresh)."
      security:
        - supabaseAccessToken: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  stripeAccountId:
                    type: string
                  url:
                    type: string
                    format: uri
                  expiresAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                required:
                  - stripeAccountId
                  - url
                  - expiresAt
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
  /v1/providers/me/stripe-connect/dashboard-link:
    post:
      summary: "Issue a Stripe Express dashboard login link — requires step-up MFA.
        Used for bank-detail edits and payout withdrawals (OH-110 AC #3, AC
        #4)."
      tags:
        - providers
      description: "Returns a one-time Stripe Express dashboard login URL. Bank-detail
        edits and payout withdrawals both happen inside the Express dashboard,
        so this single MFA-gated endpoint covers OH-110 AC #3 (bank-detail
        change requires step-up MFA) and AC #4 (withdrawal initiation requires
        step-up MFA). Requires a step-up MFA grant issued within the last 5
        minutes (`POST /v1/auth/step-up` from OH-103); without it the auth
        plugin returns 403 `step_up_required`."
      security:
        - supabaseAccessToken: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                  createdAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                required:
                  - url
                  - createdAt
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
...
servers:
  - url: https://api.ourhaven.example
    description: Production (Fly.io iad)
  - url: http://localhost:8080
    description: Local development
...